refactor(reviwer solicitação): Refatoração das solicitações do review
Adicionado texto estaticos em constantes, adicionando os styles ao padrão utilizando styled-components, separando requisição a api do arquivo de interface
