+++ src/components/IconButton/index.tsx
@@ -1,6 +1,5 @@
-import { Colors } from '@utils/Colors';
 import React, { ReactElement, ReactNode } from 'react';
-import Icon from 'react-native-vector-icons/Feather';
+import { Icon } from './styles';
 import { IconButtonProps } from 'react-native-vector-icons/Icon';
 
 interface IIconButtonProps extends IconButtonProps {
@@ -11,21 +10,12 @@ interface IIconButtonProps extends IconButtonProps {
 
 function IconButton({ updateIcon, children, name }: IIconButtonProps): ReactElement {
     return (
-        <Icon.Button
-            size={22}
-            iconStyle={{
-                flex: 1,
-                top: 24,
-                right: 20,
-                position: 'absolute'
-            }}
-            color={Colors.darkGray}
-            backgroundColor={Colors.transparent}
+        <Icon
             onPress={updateIcon}
             name={name}
         >
             {children}
-        </Icon.Button>
+        </Icon>
     );
 }
