+++ src/services/env.js
@@ -1,0 +1,1 @@
+export const baseUrl = "https://meusbeats.herokuapp.com/api";
